### 2019-05-19 #1460 @arfon
:wave: @RuthAngus - [this PR](https://github.com/RuthAngus/stardate/pull/4) should fix your paper compilation issues.
