### 2019-05-20 #1460 @arfon
Sorry, looks like you need [this fix](https://github.com/RuthAngus/stardate/pull/5) too.

@RuthAngus - does this paper accompany a submission to a AAS journal?
