### 2019-05-20 #1460 @RuthAngus
Thanks, @arfon! Yes, it does -- I submitted a full-length paper to AJ about the method!
