+++ paper/paper.md
@@ -50,7 +50,7 @@ The paper describing the method in more detail is Angus et al. (submitted).
 
 ``stardate`` is built on top of the ``isochrones`` Python package
 [@isochrones] and uses the MIST stellar evolution models [@dotter2016],
-@[choi2016].
+[@choi2016].
 Development of ``stardate`` happens on GitHub and any issues can be raised
 there.
 
