### 2019-05-20 #1460 @RuthAngus
Hi @arfon! I noticed a typo in the paper which I have fixed now!
