### 2019-05-20 33d08ee
updating docs to reflect change in isochrones install from bolo to master
