### 2019-05-21 #1460 @arfon
@RuthAngus - could you suggest some potential reviewers for this submission? [This list of people](https://bit.ly/joss-reviewers) might be a good place to start.

Also cc @crawfordsm on this submission.
