### 2019-05-21 23ba1b4
removing dynesty dependency
