### 2019-05-21 7da872f
updating docs to reflect installation
