### 2019-05-22 #1460 @nespinoza
Just wanted to say that this package seems extremely interesting and if @mattpitkin for some reason can't review this I would love to as well :-).
