### 2019-05-22 #1469 @whedon
[REVIEW]: stardate:  Combining dating methods for better stellar ages
**Submitting author:** @ruthangus (<a href="http://orcid.org/0000-0003-4540-5661">Ruth Angus</a>)
**Repository:** <a href="https://github.com/RuthAngus/stardate" target ="_blank">https://github.com/RuthAngus/stardate</a>
**Version:** v0.1.0
