### 2019-05-22 0820b91
setting max_n in example
