### 2019-05-22 #1469 @mattpitkin
Hi @RuthAngus, one of the things we're asked to check is:

> **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Would you be able to add a few sentences to the README file and main documentation page telling people how they should contribute to the project? Just something along the lines of saying that if they want to contribute they can submit issues or pull requests from the github repo.
