### 2019-05-22 #1469 @RuthAngus
Absolutely! I have added a sentence to the bottom of the README.
