### 2019-05-23 #1469 @mattpitkin
@RuthAngus thanks for adding this. I noticed that in the raw version of the README.rst file there's some info on installing and an example of running the code, but these are commented out. Even with the link to the documentation that exists I think it would still be useful to have that info visible. Could you uncomment them?
