### 2019-05-23 b10ff1c
adding install and example to README
