+++ README.rst
@@ -133,48 +133,3 @@ M dwarfs on the main sequence, older than a few hundred Myrs.
 If you're not sure if gyrochronology is applicable to your star, want the best
 age possible, or would like proper uncertainty estimates, I recommend using
 the full MCMC approach.
-
-
-.. You'll also need to download isochrones and switch to the eep branch:
-
-.. .. code-block:: bash
-..     git clone https://github.com/timothydmorton/isochrones
-..     cd isochrones
-..     git checkout eep
-..     python setup.py install
-
-.. In order to get started you can create a dictionary containing the observables
-.. you have for your star.
-.. These could be atmospheric parameters (like those shown in the example below
-.. for the Sun), or just photometric colors, like those from *2MASS*, *SDSS* or
-.. *Gaia*.
-.. If you have a parallax, asteroseismic parameters, or an idea of the
-.. maximum V-band extinction you should throw those in too.
-.. Set up the star object and :func:`chronology.star.fit` will run Markov Chain
-.. Monte Carlo (using *emcee*) in order to infer a Bayesian age for your star.
-
-.. Example usage
-.. =============
-.. ::
-
-..     import stardate as sd
-
-..     # Create a dictionary of observables
-..     iso_params = {"teff": (5777, 10),     # Teff with uncertainty.
-..                   "logg": (4.44, .05),    # logg with uncertainty.
-..                   "feh": (0., .001),      # Metallicity with uncertainty.
-..                   "parallax": (1., .01),  # Parallax in milliarcseconds.
-..                   "maxAV": .1}            # Maximum extinction
-
-..     prot, prot_err = 26, 1
-
-..     # Set up the star object.
-..     star = sd.star(iso_params, prot, prot_err)  # Here's where you add a rotation period
-
-..     # Run the MCMC
-..     star.fit()
-
-..     # Print the median age with the 16th and 84th percentile uncertainties.
-..     print("stellar age = {0} + {1} + {2}".format(star.age[0], star.age[2], star.age[1])
-
-..     >> stellar age = 4.5 + 2.1 - 1.3
