### 2019-05-25 #1469 @mattpitkin
@RuthAngus I'm working through the very nice example here https://stardate.readthedocs.io/en/latest/tutorials/Tutorial.html#. When I get to the code block after the sentence `"If you know the extinction you can pass that to the Star object too:"` it throws an error because `Av` and `Av_err` are not defined. I'm assuming this `Av` is the extinction value (and it's associated error) that previous sentence mentions, but this should be made explicit in the sentence, or the values should be removed from the example, so that it will run without an error if people have been following step-by-step.
