### 2019-05-26 #1469 @mattpitkin
@RuthAngus Could you add `emcee` to the dependencies listed in https://stardate.readthedocs.io/en/latest/user/install.html#dependencies?
