### 2019-05-26 #1469 @mattpitkin
@RuthAngus In the paper it's probably worth mentioning that the fitting uses the MCMC software `emcee`  to perform the fitting (with a reference) and therefore returns samples from the parameters' posterior probability distributions.

Related to this, is there any way to set what priors you can use for any of the fitted parameters?
