### 2019-05-26 #1469 @mattpitkin
@RuthAngus are there any plans to add a test suite and continuous integration to the package? I don't think this is a requirement for software that is published in JOSS (@arfon?), but it would be nice and might be useful for future development?
