### 2019-05-26 #1469 @arfon
> @RuthAngus are there any plans to add a test suite and continuous integration to the package? I don't think this is a requirement for software that is published in JOSS (@arfon?), but it would be nice and might be useful for future development?

The [JOSS guidelines on testing](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) require at _least_ the following:

> Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g. a sample input file to assert behaviour)

We definitely prefer automated tests (ideally hooked up to Travis/Circle CI etc too) but don't require them. It is required though that there are documented procedures that the reviewer can follow to _objectively_ check that the software is doing what is expected.

@mattpitkin - does that help?
