### 2019-05-26 #1469 @mattpitkin
@arfon there is a very nice example and tutorial of using the code, but I think from what you say this may not quite be enough. As the code is doing stochastic sampling, the example given won't necessarily produce identical results each time.

@RuthAngus here's what I'd suggest by way of a test, which may be useful. First, make the `star` class, or its `run_mcmc()` method accept a random state variable of the form required by [emcee](https://emcee.readthedocs.io/en/stable/api.html#emcee.Sampler.random_state). With this you'll be able to seed the sampler to reproduce exactly the same results each time. Then run the code for a particular star that you've analysed before, using a random state initialised with a particular seed number, and probably starting the sampler very close the already known maximum posterior values, generating a few hundred samples. Save these samples and add then into the repo in a `test` directory. Also, add the script that you've just used to produce the samples to the `test` directory, but change it so that the samples get output to a different file. Also, have the script read in the previously generated samples file and have check that the new samples it will produce are identical to the already existing samples (you could compare each sample, or just compare statistics like the mean, median and credible intervals to make sure they are the same). Does that sound doable?
