### 2019-05-29 fa37710
adding emcee dependency
