### 2019-05-29 9f4ca62
adding emcee citation
