+++ paper/references.bib
@@ -11,6 +11,23 @@ archivePrefix = "ascl",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{emcee,
+   author = {{Foreman-Mackey}, D. and {Hogg}, D.~W. and {Lang}, D. and {Goodman}, J.
+	},
+    title = "{emcee: The MCMC Hammer}",
+  journal = {\pasp},
+archivePrefix = "arXiv",
+   eprint = {1202.3665},
+ primaryClass = "astro-ph.IM",
+     year = 2013,
+    month = mar,
+   volume = 125,
+    pages = {306},
+      doi = {10.1086/670067},
+   adsurl = {http://adsabs.harvard.edu/abs/2013PASP..125..306F},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{dotter2016,
    author = {{Dotter}, A.},
     title = "{MESA Isochrones and Stellar Tracks (MIST) 0: Methods for the Construction of Stellar Isochrones}",
