+++ stardate/tests/integration_tests.py
@@ -12,64 +12,64 @@ mist = get_ichrone("mist", bands=bands)
 import corner
 
 
-def test_sim():
-    full_df = pd.read_csv("../../paper/code/data/simulated_data.csv")
+# def test_sim():
+#     full_df = pd.read_csv("../../paper/code/data/simulated_data.csv")
 
-    i = 15
-    df = full_df.iloc[i]
+#     i = 15
+#     df = full_df.iloc[i]
 
-    teff_err, logg_err, feh_err = 25, .05, .05
+#     teff_err, logg_err, feh_err = 25, .05, .05
 
-    iso_params = {"teff": (df["teff"], teff_err),
-                 "logg": (df["logg"], logg_err),
-                 "feh": (df["feh"], feh_err),
-                 "J": (df["jmag"], df["J_err"]),
-                 "H": (df["hmag"], df["H_err"]),
-                 "K": (df["kmag"], df["K_err"]),
-                 "B": (df["B"], df["B_err"]),
-                 "V": (df["B"], df["V_err"]),
-                 "G": (df["G"], df["G_err"]),
-                 "BP": (df["BP"], df["BP_err"]),
-                 "RP": (df["RP"], df["RP_err"]),
-                 "parallax": (df["parallax"], df["parallax_err"]),
-                 "maxAV": .1}
-    print(df["parallax"], df["parallax_err"])
+#     iso_params = {"teff": (df["teff"], teff_err),
+#                  "logg": (df["logg"], logg_err),
+#                  "feh": (df["feh"], feh_err),
+#                  "J": (df["jmag"], df["J_err"]),
+#                  "H": (df["hmag"], df["H_err"]),
+#                  "K": (df["kmag"], df["K_err"]),
+#                  "B": (df["B"], df["B_err"]),
+#                  "V": (df["B"], df["V_err"]),
+#                  "G": (df["G"], df["G_err"]),
+#                  "BP": (df["BP"], df["BP_err"]),
+#                  "RP": (df["RP"], df["RP_err"]),
+#                  "parallax": (df["parallax"], df["parallax_err"]),
+#                  "maxAV": .1}
+#     print(df["parallax"], df["parallax_err"])
 
-    mod = StarModel(mist, **iso_params)
-    args = [mod, df["prot_praesepe"], df["prot_err_praesepe"], False, False,
-            True, "praesepe"]
-    inits=[329.58, 9.5596, -.0478, np.log(260), .0045]
-    print(lnprob(inits, *args))
-    assert 0
+#     mod = StarModel(mist, **iso_params)
+#     args = [mod, df["prot_praesepe"], df["prot_err_praesepe"], False, False,
+#             True, "praesepe"]
+#     inits=[329.58, 9.5596, -.0478, np.log(260), .0045]
+#     print(lnprob(inits, *args))
+#     assert 0
 
-    # lnparams = [df.eep, df.age, df.feh, np.log(df.d_kpc*1e3), df.Av]
-    # print(lnprob(lnparams, *args))
-    filename = "{}_M_test".format(str(int(df["ID"])).zfill(4))
-    star = sd.Star(iso_params, prot=df["prot"], prot_err=.01,
-                   filename=filename)
-    star.fit(max_n=20000, model="praesepe")
+#     # lnparams = [df.eep, df.age, df.feh, np.log(df.d_kpc*1e3), df.Av]
+#     # print(lnprob(lnparams, *args))
+#     filename = "{}_M_test".format(str(int(df["ID"])).zfill(4))
+#     star = sd.Star(iso_params, prot=df["prot"], prot_err=.01,
+#                    filename=filename)
[…]
