+++ stardate/tests/lhf_test.py
@@ -86,94 +86,94 @@ def test_lnprob_higher_likelihood_real():
     good_vs_bad(good_lnprob, good_lnparams, args, 10)
 
 
-def test_for_nans():
-    """
-    Something is causing the lhf to return NaN. Get to the bottom of it!
-    """
-    df = pd.read_csv("../paper/code/data/simulated_data.csv")
-    for i in range(len(df)):
-        print(i, "of", len(df))
-        iso_params = pd.DataFrame(dict({"teff": (df.teff[i], 10),
-                                "logg": (df.logg[i], .05),
-                                "feh": (df.feh[i], .001),
-                                "parallax": (df.parallax[i], .01)}))  # mas
+# def test_for_nans():
+#     """
+#     Something is causing the lhf to return NaN. Get to the bottom of it!
+#     """
+#     df = pd.read_csv("../paper/code/data/simulated_data.csv")
+#     for i in range(len(df)):
+#         print(i, "of", len(df))
+#         iso_params = pd.DataFrame(dict({"teff": (df.teff[i], 10),
+#                                 "logg": (df.logg[i], .05),
+#                                 "feh": (df.feh[i], .001),
+#                                 "parallax": (df.parallax[i], .01)}))  # mas
 
-        mod = StarModel(mist, **iso_params)
+#         mod = StarModel(mist, **iso_params)
 
-        N = 10000
-        eeps = np.random.uniform(-100, 2000, N)
-        lnages = np.random.uniform(0, 11, N)
-        fehs = np.random.uniform(-5, 5, N)
-        Ds = np.log(np.random.uniform(0, 10000, N))
-        Avs = np.random.uniform(-.2, 1.2, N)
-        # periods = 10**np.random.uniform(-1, 3, N)
-        probs, priors = [np.empty(N) for i in range(2)]
+#         N = 10000
+#         eeps = np.random.uniform(-100, 2000, N)
+#         lnages = np.random.uniform(0, 11, N)
+#         fehs = np.random.uniform(-5, 5, N)
+#         Ds = np.log(np.random.uniform(0, 10000, N))
+#         Avs = np.random.uniform(-.2, 1.2, N)
+#         # periods = 10**np.random.uniform(-1, 3, N)
+#         probs, priors = [np.empty(N) for i in range(2)]
 
-        for j in trange(N):
-            lnparams = [eeps[j], lnages[j], fehs[j], Ds[j], Avs[j]]
-            args = [mod, df.prot[i], 1., False, False, True,
-                    "angus15"]
-            probs[j] = lnprob(lnparams, *args)[0]
-            priors[j] = lnprob(lnparams, *args)[1]
+#         for j in trange(N):
+#             lnparams = [eeps[j], lnages[j], fehs[j], Ds[j], Avs[j]]
+#             args = [mod, df.prot[i], 1., False, False, True,
+#                     "angus15"]
+#             probs[j] = lnprob(lnparams, *args)[0]
+#             priors[j] = lnprob(lnparams, *args)[1]
 
-        print(len(probs), len(probs[np.isnan(probs)]))
-        assert sum(np.isnan(probs)) == 0
+#         print(len(probs), len(probs[np.isnan(probs)]))
+#         assert sum(np.isnan(probs)) == 0
 
 
-def test_likelihood_rotation_giant():
-    """
-    Make sure that the lhf can cope with zeros, NaNs and None values for the
-    rotation period.
-    Also, check that there is a drop in likelihood at the eep = 454 boundary.
-    """
-    iso_params = pd.DataFrame(dict({"teff": (5777, 10),
[…]
