### 2019-05-29 684e6c0
adding a seed argument to fit function
