### 2019-05-29 #1469 @RuthAngus
Hi @mattpitkin, thanks for these comments!

I have now fixed the Av bug in the tutorial and added emcee to the list of dependencies. Thanks for running through the tutorial! There is not currently any support for manually adjusting the priors, but that's a feature I'd love to add in future. Would you mind if I tabled this for the next version? The priors that I'm using are described in the appendix of the submitted AAS paper that accompanies this software. I have also added a mention and reference to emcee in the JOSS paper, thanks for suggesting that!

I do currently have some tests in the /stardate/tests directory but I don't provide instructions on how to run these. Should I add some instructions to the docs on how to run the tests with pytest or something? The test in integration_tests.py does something similar to what you suggest. It does use a fixed random number seed, but I had previously hard-coded this into the function that runs emcee, which doesn't seem like the most sensible choice :-). I have now added an argument to set the seed in the fit() function. This test doesn't save samples like you suggest, which is a great idea, but that would fail each time I changed something in the likelihood function or prior, so I'd need to regenerate the sample file each time I did that. Do you think this is worthwhile doing regardless? I also agree that adding continuous integration would be really nice. I've never set this up before and I'm not that familiar with how, e.g. Travis works, but would it be a problem that the big file of MIST model isochrones needs to be downloaded in order to run the tests?
