### 2019-05-30 #1469 @mattpitkin
@RuthAngus here are some replies inline.

> I have now fixed the Av bug in the tutorial and added emcee to the list of dependencies.

Great, thanks. I think emcee is still missing in the installation requirements described in the README file, but otherwise its looking good.

> I do currently have some tests in the /stardate/tests directory but I don't provide instructions on how to run these. Should I add some instructions to the docs on how to run the tests with pytest or something? The test in integration_tests.py does something similar to what you suggest.

Excellent. I'd missed that this test was there, but it looks like the sort of thing that is required. Adding instructions on how to run it with pytest would be very useful.

> It does use a fixed random number seed, but I had previously hard-coded this into the function that runs emcee, which doesn't seem like the most sensible choice :-). I have now added an argument to set the seed in the fit() function.

Great :smiley:

> This test doesn't save samples like you suggest, which is a great idea, but that would fail each time I changed something in the likelihood function or prior, so I'd need to regenerate the sample file each time I did that.

My suggestion of having a saved set of samples in the test suite and doing a direct comparison between those and a new set of samples may be overkill. Having a broader consistency test like you do in your current `integration_tests.py` should be enough (and is enough for this review I think). However, in general it would be good to have some fixed exact comparison - as you say, you may change the likelihood/prior in the future, but when doing this it would be best to fully maintain backwards compatibility (unless you find a bug!), and a fixed test for a specific set of data and a specific likelihood function and prior, would help ensure that this backwards compatibility is maintained. I don't think this is a requirement for this review though.

> I also agree that adding continuous integration would be really nice. I've never set this up before and I'm not that familiar with how, e.g. Travis works, but would it be a problem that the big file of MIST model isochrones needs to be downloaded in order to run the tests?

It's relatively easy to set up Travis CI integration - an example of a `.travis.yml` file that I've used in one of my packages for this is [here](https://github.com/mattpitkin/psrqpy/blob/master/.travis.yml). Once you have that in your repo you just have to register the project on the [travis](https://travis-ci.org/) website. It think is should be able to handle having to download the MIST model isochrones file each time - often it has to download a large Docker image file, or install some large package, each time anyway (in fact you could create your own Docker image for it to use that already includes the file...)
