### 2019-05-30 ea1509a
adding test instructions
