+++ requirements.txt
@@ -1,0 +1,7 @@
+numpy
+pandas
+h5py
+numba
+emcee
+tqdm
+isochrones
