+++ README.rst
@@ -45,20 +45,21 @@ The dependencies of *stardate* are
 `pandas <https://pandas.pydata.org/>`_,
 `h5py <https://www.h5py.org/>`_,
 `numba <http://numba.pydata.org/>`_,
+`emcee <https://emcee.readthedocs.io/en/stable/`_,
 `tqdm <https://tqdm.github.io/>`_ and
 `isochrones <https://github.com/timothydmorton/isochrones>`_.
 
-The first five of these can be installed using conda or pip:
+The first six of these can be installed using conda or pip:
 
 .. code-block:: bash
 
-    conda install numpy pandas h5py numba tqdm
+    conda install numpy pandas h5py numba emcee tqdm
 
 or
 
 .. code-block:: bash
 
-    pip install numpy pandas h5py numba tqdm
+    pip install numpy pandas h5py numba emcee tqdm
 
 You'll also need to download isochrones:
 
