+++ docs/source/user/tests.rst
@@ -1,12 +1,10 @@
 Running tests
 ============
 
-You can test *stardate* by navigating to the /stardate/tests directory and
-either running the test files directly:
+You can test *stardate* using `pytest <https://docs.pytest.org/en/latest/>`_
+from the code base.
 
 .. code-block:: bash
 
-    python lhf_test.py
-    python integration_tests.py
-
-or using `pytest <https://docs.pytest.org/en/latest/>`_ to run the test suite.
+    pip install pytest
+    pytest
