### 2019-05-30 1c69ac9
fixing for pytest
