### 2019-05-30 7eb4317
adding Travis badge
