+++ docs/source/user/tests.rst
@@ -1,10 +1,11 @@
 Running tests
 ============
 
-You can test *stardate* from the code base using `pytest
+You can test *stardate* from within the base directory using `pytest
 <https://docs.pytest.org/en/latest/>`_.
 
 .. code-block:: bash
 
     pip install pytest
+    cd stardate
     pytest
