+++ docs/source/user/tests.rst
@@ -3,9 +3,9 @@ Running tests
 
 You can test *stardate* from within the base directory using `pytest
 <https://docs.pytest.org/en/latest/>`_.
+Install pytest using pip if you don't already have it installed, then navigate
+to the *stardate* base directory and type:
 
 .. code-block:: bash
 
-    pip install pytest
-    cd stardate
     pytest
