+++ .travis.yml
@@ -12,7 +12,7 @@ install:
   - pip install --upgrade pip
   - pip install -r requirements.txt
   # install packages for documentation building
-  - pip install sphinx sphinx-rtd-theme
+  # - pip install sphinx sphinx-rtd-theme
   # install packages for testing
   - pip install --upgrade pytest>=3.6.3
   # build stardate
