+++ README.rst
@@ -15,7 +15,8 @@ stardate
 .. image:: https://travis-ci.org/RuthAngus/stardate.svg?branch=master
     :target: https://travis-ci.org/RuthAngus/stardate
 
-
+.. image:: https://readthedocs.org/projects/stardate/badge/?version=latest
+:target: https://stardate.readthedocs.io/en/latest/?badge=latest
 
 Checkout `the documentation <https://stardate.readthedocs.io/en/latest/>`_.
 
