### 2019-05-30 a5da54d
did I break things by adding a new requirements.txt? Removing that and installing 1-by-1
