+++ .travis.yml
@@ -10,7 +10,13 @@ python:
   - "3.7"
 install:
   - pip install --upgrade pip
-  - pip install -r requirements.txt
+  # - pip install -r requirements.txt
+  - pip install numpy
+  - pip install pandas
+  - pip install h5py
+  - pip install numba
+  - pip install tqdm
+  - pip install isochrones
   # install packages for documentation building
   # - pip install sphinx sphinx-rtd-theme
   # install packages for testing
