### 2019-05-30 36e4f6d
switching back to requirements.txt
