+++ docs/source/user/tests.rst
@@ -1,10 +1,11 @@
 Running tests
 ============
 
-You can test *stardate* using `pytest <https://docs.pytest.org/en/latest/>`_
-from the code base.
+You can test *stardate* from within the base directory using `pytest
+<https://docs.pytest.org/en/latest/>`_.
+Install pytest using pip if you don't already have it installed, then navigate
+to the *stardate* base directory and type:
 
 .. code-block:: bash
 
-    pip install pytest
     pytest
