### 2019-05-30 aa364a3
Update .travis.yml
