### 2019-05-30 2d514f1
Update readthedocs.yml
