+++ readthedocs.yml
@@ -1,4 +1,4 @@
 requirements_file: docs/requirements.txt
 python:
     version: 3.5
-    pip_install: true
+    # pip_install: true
