### 2019-05-30 9dd0cbf
changes to installation and more tutorial
