### 2019-05-31 #1469 @RuthAngus
@mattpitkin I've added emcee to the README installation guide now, thanks for spotting that. I have also added some [basic test instructions](https://stardate.readthedocs.io/en/latest/user/tests.html) to the docs. Does this look alright to you? Thanks for providing an example of a .travis.yml! I have CI working now and I'm really pleased about this! Let me know if you have any other suggestions for improvements.
