### 2019-05-31 #1469 @nespinoza
Hi @RuthAngus,

 Thanks first of all for all the work you've already done with @mattpitkin's part of the review --- I've been keeping an eye on his comments in order to not bring up the same ideas/comments again in my review.

 There's a couple of things that I think should be stated in the `stardate` documentation with which I actually had trouble with; this is what prevents me right now from ticking the **Installation instructions** part, but I think they are fairly easy to address:

1. Because `isochrones` only works now on `python3`, `stardate` also only works on `python3`. As a hard-core `python2.7` user (and yes, I know --- we have to evolve! On it), I think it is important that you state at the beginning of the docs that `stardate` will only work with `python3` because of this reason.

2. On a similar note, it appears that when you do a `pip install emcee`, in several systems I have tried this the default is to install emcee 2.2.1, which _does not include the `emcee.backend` module_ (which is only introduced in emcee version 3), which is needed for `stardate` to work (line 176 in `stardate.py`). I think this really should be a shout-out to @dfm, but in the meantime perhaps it should be stated that `stardate` also only works on `emcee` versions 3 and up in the docs as well.

Other than the above, I believe this is a fantastic package, kudos to @RuthAngus et al. for it --- looking forward to using it in my own research!
