### 2019-05-31 #1469 @RuthAngus
Hi @nespinoza, thanks for these suggestions -- these are both things I hadn't thought of! I have now added a python3 warning to the readme and [docs](https://stardate.readthedocs.io/en/latest/index.html) and I have made the emcee version requirement explicit in the [installation instructions](https://stardate.readthedocs.io/en/latest/user/install.html). Hopefully that should work now!
