### 2019-05-31 #1469 @nespinoza
Fantastic @RuthAngus, thank you for including that. I have signed the **Installation Instructions** part then and thus, I'm happy to recommend this por publication in JOSS!
