### 2019-05-31 91b3654
fixing install typo
