### 2019-05-31 #1469 @arfon
> Thanks so much @nespinoza and @mattpitkin for helping me improve the code! @arfon, is there anything else I need to do now?

At this point I think we need to wait for your AAS paper to be accepted so we can publish them at the same time. Do you know the status of that? Once that paper is accepted too then we can make a final update to the JOSS paper to make sure that your `Angus et al (submitted)` can be updated to cite the accepted article (including DOI etc).

/ cc @crawfordsm
