### 2019-05-31 #1469 @RuthAngus
@arfon okay, sounds good! I'm currently waiting to receive a first-round referee report for the AAS paper. I submitted it less than two weeks ago, so it could still be quite a while before getting an acceptance for that one. I'm in no rush though and happy to wait :-)

/ cc @crawfordsm
