### 2019-06-10 bbf720e
testing out lower Ro threshold
