### 2019-06-11 b6f5436
test fix to work with new model function
