### 2019-06-13 ec09531
relaxing the EEP prior
