### 2019-06-13 ccfaa56
Ro_cutoff = 2
