### 2019-06-14 aee841d
rossby cut back to 1.5
