+++ stardate/lhf.py
@@ -594,9 +594,9 @@ def sigma(eep, log_age, feh, color, model="praesepe"):
             B-V color if model == "angus15"
 
     """
-    kcool, khot, keep = 100, 100, .2
+    kcool, khot, keep = 100, 100, 10# .2
     Lcool, Lhot, Leep = .5, .5, 5
-    x0eep = 400
+    x0eep = 410
     k_old, x0_old = 100, np.log10(10*1e9)
     k_young, x0_young = 20, np.log10(250*1e6)
     L_age = .5
