### 2019-06-14 fad3f33
eep 380 with less smooth transition
