### 2019-06-27 5377a20
SingleStarModel rather than StarModel
