### 2019-07-01 45f34e5
adding save_samples as a kwarg
