### 2019-07-01 5b2d534
default number of walkers is now 50 (and the last commit included some changes to initialization)
