+++ test/lhf_test.py
@@ -86,96 +86,6 @@ def test_lnprob_higher_likelihood_real():
     good_vs_bad(good_lnprob, good_lnparams, args, 10)
 
 
-# def test_for_nans():
-#     """
-#     Something is causing the lhf to return NaN. Get to the bottom of it!
-#     """
-#     df = pd.read_csv("paper/code/data/simulated_data.csv")
-#     for i in range(len(df)):
-#         print(i, "of", len(df))
-#         iso_params = pd.DataFrame(dict({"teff": (df.teff[i], 10),
-#                                 "logg": (df.logg[i], .05),
-#                                 "feh": (df.feh[i], .001),
-#                                 "parallax": (df.parallax[i], .01)}))  # mas
-
-#         mod = StarModel(mist, **iso_params)
-
-#         N = 10000
-#         eeps = np.random.uniform(-100, 2000, N)
-#         lnages = np.random.uniform(0, 11, N)
-#         fehs = np.random.uniform(-5, 5, N)
-#         Ds = np.log(np.random.uniform(0, 10000, N))
-#         Avs = np.random.uniform(-.2, 1.2, N)
-#         # periods = 10**np.random.uniform(-1, 3, N)
-#         probs, priors = [np.empty(N) for i in range(2)]
-
-#         for j in trange(N):
-#             lnparams = [eeps[j], lnages[j], fehs[j], Ds[j], Avs[j]]
-#             args = [mod, df.prot[i], 1., False, False, True,
-#                     "angus15"]
-#             probs[j] = lnprob(lnparams, *args)[0]
-#             priors[j] = lnprob(lnparams, *args)[1]
-
-#         print(len(probs), len(probs[np.isnan(probs)]))
-#         assert sum(np.isnan(probs)) == 0
-
-
-# def test_likelihood_rotation_giant():
-#     """
-#     Make sure that the lhf can cope with zeros, NaNs and None values for the
-#     rotation period.
-#     Also, check that there is a drop in likelihood at the eep = 454 boundary.
-#     """
-#     iso_params = pd.DataFrame(dict({"teff": (5777, 10),
-#                                 "logg": (4.44, .05),
-#                                 "feh": (0., .001),
-#                                 "parallax": (1., .01)}))  # mas
-#     # Set up the StarModel isochrones object.
-#     mod = StarModel(mist, **iso_params)
-#     lnparams = [355, np.log10(4.56*1e9), 0., np.log(1000), 0.]
-
-#     args = [mod, None, None, False, False, True, "angus15"]  # the lnprob arguments]
-#     none_lnprob = lnprob(lnparams, *args)
-
-#     args = [mod, np.nan, np.nan, False, False, True, "angus15"]  # the lnprob arguments]
-#     nan_lnprob = lnprob(lnparams, *args)
-
-#     args = [mod, 0., 0., False, False, True, "angus15"]  # the lnprob arguments]
-#     zero_lnprob = lnprob(lnparams, *args)
-
-#     args = [mod, 26., 1., True, False, True, "angus15"]  # the lnprob arguments]
-#     iso_lnprob = lnprob(lnparams, *args)
-
-#     # check that gyro is switched off for all of these.
-#     none_lnprob == nan_lnprob
-#     nan_lnprob == zero_lnprob
-
-#     # Likelihood should be greater for dwarfs because gyro lnlike is a broad
-#     # Gaussian for giants.
-#     giant_params = [455, np.log10(4.56*1e9), 0., np.log(1000), 0.]
[…]
