### 2019-07-16 349225d
adding emcee version to requirements
