+++ paper/method.tex
@@ -56,6 +56,12 @@ line in figure \ref{fig:praesepe}).
 In order to fit a relation to Praesepe, we removed rotational outliers bluer
 than \gcolor\ = 2.7 via sigma-clipping and fit a 5th-order polynomial to the
 remaining FGK and early M stars.
+\racomment{We found that a 5th order polynomial provided a substantially
+better fit than lower-order polynomials, which were not able to capture the
+sharp `elbow’ in the rotation period-color relation. Additional orders
+provided either a worse
+fit, tending towards extreme values at the boundaries, or diminishing returns
+in goodness-of-fit.}
 We also fit a straight line to the late M dwarfs (\gcolor\ $>$ 2.7), to
 capture the mass-dependent initial rotation periods of low mass stars
 \citep{somers2017}.
@@ -66,19 +72,19 @@ performed in a Jupyter notebook available at
 This new Praesepe-calibrated gyrochronology relation is,
 \begin{equation}
     \log_{10}(P_\mathrm{rot}) =
-    c_A\log_{10}(A) +
+    c_A\log_{10}(t) +
     \sum_{n=0}^4 c_n[\log_{10}(G_{BP}-G_{RP})]^n
 \label{eqn:fgk_gyro}
 \end{equation}
 for stars with \gcolor\ $<$ 2.7 and
 \begin{equation}
     \log_{10}(P_\mathrm{rot}) =
-    c_A\log_{10}(A) +
+    c_A\log_{10}(t) +
     \sum_{m=0}^1 b_m[\log_{10}(G_{BP}-G_{RP})]^m
 \label{eqn:m_gyro}
 \end{equation}
 for stars with \gcolor\ $>$ 2.7, where $P_{\mathrm{rot}}$ is rotation period
-in days and $A$ is age in years.
+in days and $t$ is age in years.
 Best-fit coefficient values are shown in table \ref{tab:coefficients}.
 \begin{table}[h!]
   \begin{center}
@@ -199,17 +205,17 @@ The rotation period model for each of these groups is described below.
         This is useful because we have not yet built an accurate rotation-age
         relation for subgiants into our model.}
     \item{We model stars younger than around 250 Myrs with a log-normal
-        distribution with mean function given by \ref{eqn:fgk_gyro},
+        distribution with mean function given by equation \ref{eqn:fgk_gyro},
         \ref{eqn:m_gyro} or 0.56, depending on it whether it has
-        an intermediate, blue or red color, respectively and a inflated
+        an intermediate, red, or blue color respectively and a inflated
         standard deviation of 0.5 dex.
         Rotation periods in young open clusters show a large amount of scatter
         because they have not yet converged onto the \citet{skumanich1972}
         spin-down sequence.}
-    \item{Finally, we model stars with very low and high metallicites (-3 $>$
-        [Fe/H] $>$ 3) with a log-normal distribution with mean given by
-        \ref{eqn:fgk_gyro}, \ref{eqn:m_gyro} or 0.56, depending on its color,
-        and a standard deviation of 0.5 dex.
+    \item{Finally, we model stars with very low and high metallicites (-0.2
+        $>$ [Fe/H] $>$ 0.2) with a log-normal distribution with mean given by
+        equation \ref{eqn:fgk_gyro}, \ref{eqn:m_gyro} or 0.56, depending on
[…]
