+++ paper/intro.tex
@@ -14,7 +14,7 @@ On the MS, isochrones are tightly spaced and, even with very precise
 measurements of effective temperature and luminosity, the position of a MS
 star on the HRD may be consistent with range of isochrones spanning several
 billion years.
-At main sequence turn-off however, isochrones are spread further apart, so
+At the main-sequence turnoff however, isochrones are spread further apart, so
 that sufficiently precisely measured temperatures and luminosities can yield
 ages that are extremely precise (with minimum statistical uncertainties on the
 order of 5-10\%) \citep[\eg][]{pont2004}.
@@ -152,7 +152,7 @@ new Praesepe-based model, which is built into an open source {\it Python}
 package for gyrochronology called \sd.
 \sd\ provides the framework for simultaneous gyrochronology and isochrone
 fitting, and, because \sd\ is modular, it would be straightforward to update
-the gyrochronology relation in future.
+the gyrochronology relation in the future.
 
 This paper is laid out as follows.
 In section \ref{section:method} we describe our new age-dating model and its
@@ -161,10 +161,10 @@ simulated stars and cluster stars, and in section \ref{section:conclusion} we
 discuss the implications of these tests and future pathways for development.
 Throughout this paper we use the term {\it `observables'} to refer to the
 following observed properties of a star: \teff, \logg, observed bulk
-metallicity ($[\hat{\mathrm{Fe/H}}]$), parallax ($\bar{\omega}$), photometric
+metallicity ($[\mathrm{Fe/H}]$), parallax ($\bar{\pi}$), photometric
 colors in different passbands (${\bf m_x} = [m_J, m_H, m_K, m_B, m_V, m_G,
 m_{GBP}, m_{GRP}...]$, etc) and rotation period ($P_{\mathrm{rot}}$).
 The term `parameters' refers to the physical properties of that star: age
-($A$), equivalent evolutionary phase (EEP), metallicity ($\mathrm{Fe/H}$),
-distance ($D$) and V-band extinction ($A_V$).
+($t$), equivalent evolutionary phase (EEP), model metallicity
+($\mathrm{M/H}$), distance ($D$) and V-band extinction ($A_V$).
 These are the properties that generate the observables.
