+++ paper/ms.tex
@@ -42,7 +42,7 @@
 \newcommand{\FeH}{[Fe/H]}
 \newcommand{\feh}{[Fe/H]}
 \newcommand{\prot}{$P_{\mathrm{rot}}$}
-\newcommand{\pmega}{$\bar{\omega}$}
+\newcommand{\pmega}{$\bar{\pi}$}
 \newcommand{\mj}{$m_j$}
 \newcommand{\mh}{$m_h$}
 \newcommand{\mk}{$m_k$}
