+++ paper/acknowledgements.tex
@@ -17,6 +17,8 @@ Data Processing and Analysis Consortium (DPAC,
 has been provided by national institutions, in particular the institutions
 participating in the {\it Gaia} Multilateral Agreement.
 
+This research was supported in part by the National Science Foundation under
+Grant No. NSF PHY-1748958.
 
 % \software{Astropy \citep{http://dx.doi.org/10.1051/0004-6361/201322068},
 %           Matplotlib \citep{http://dx.doi.org/10.1109/MCSE.2007.55}},
