### 2019-08-08 523876c
adding comments from Jason and Rocio
