+++ paper/method.tex
@@ -19,7 +19,7 @@ members of the 650 Myr Praesepe cluster \citep{fossati2008} by identifying
 Praesepe members with measured rotation periods from \citet{douglas2017} in
 the \ktwo-\gaia\ crossmatch catalog provided at
 \url{https://gaia-kepler.fun/}.
-This catalog cross-matched the EPIC catalog \citep{huber2016} with with the
+This catalog cross-matched the EPIC catalog \citep{huber2016} with the
 Gaia DR2 catalog \citep{brown2018}, using a 1'' search radius.
 The result was a sample of 757 stars with rotation periods, parallaxes and
 \gaia\ $G$, $G_{BP}$ and $G_{RP}$-band photometry, shown in figure
