+++ paper/conclusion.tex
@@ -16,8 +16,7 @@ We fit a new broken power law gyrochronology model to the Praesepe cluster
 and included a modification recommended by \citet{vansaders2016} that accounts
 for weakened magnetic braking at Rossby numbers larger than 2.
 The rotation periods of hot stars, cool stars, evolved stars, young stars and
-extremely metal poor and rich stars were modeled with a broad log-normal
-distribution.
+metal poor and rich stars were modeled with a broad log-normal distribution.
 We tested \sd\ on simulated data and cluster stars and demonstrated that
 combining gyrochronology with isochrone fitting improves the precision of age
 estimates for FGK dwarfs by a factor of 3 over isochrone fitting alone,
@@ -36,15 +35,15 @@ We also tested \sd\ on a set of 21 \kepler\ asteroseismic stars
 We found that discrepancies between ages measured with \sd\ and ages measured
 with asteroseismology are likely produced by differences between the MIST and
 BaSTI stellar evolution models.
-In general, the new Praesepe-based gyrochronology relation is not a good model
-for all asteroseismic stars.
-Ultimately, these empirical gyrochronology relations need further, and
-more detailed calibration.
+% In general, the new Praesepe-based gyrochronology relation is not a good model
+% for all asteroseismic stars.
+% Ultimately, these empirical gyrochronology relations need further, and
+% more detailed calibration.
 Asteroseismic and cluster stars provide an opportunity for calibration but
 given the high-dimensionality of the gyrochronology relations (\ie\ rotation
 period depends on age, mass, metallicity, surface gravity, etc), many stars
-with precise ages, spanning a range of properties, are still needed to {\it
-reliably} calibrate them.
+with precise ages, spanning a range of properties, are still needed to
+reliably calibrate them.
 }
 
 In cases where gyrochronology predicts inaccurate stellar ages it is either
