+++ paper/results.tex
@@ -422,11 +422,11 @@ measurements of G and K star rotation periods is a function of period, $\Delta
 This variation is presumably caused by a latitudinal drift in the dominant
 active regions, which traces the stellar cycle over several years, in
 combination with latitudinal differential rotation.
-This suggests that latitudinal spot drifting does not affect significantly
+This suggests that latitudinal spot drifting does not significantly
 affect stellar rotation periods when stars are young, for example the scatter
 of rotation periods about the mean gyrochronology model in Praesepe is only
 around 5\%.
-However it is likely that this effect is will become more important at older
+However it is likely that this effect will become more important at older
 ages.
 % The impact of this phenomenon on gyrochronology remain to be seen, and an
 % in-depth study is warranted.
@@ -436,6 +436,7 @@ gyrochronology is key to understanding the power of gyrochronology as an
 age-dating method.
 For now, we leave this exploration for a future study.}
 
+\subsection{Test 3: Kepler asteroseismic stars}
 \racomment{In order to test our method in the regime where both isochrone
 fitting and gyrochronology become important, we recovered the ages of the 21
 asteroseismic stars analysed in \citet{vansaders2016}.
@@ -535,8 +536,11 @@ for its mass and age, even when weakened braking is taken into account.
 This star is around Solar mass (1.0 $\pm$ 0.03 M$_\odot$) and has an
 asteroseismic age older than the Sun (7.28 $\pm$ 0.51 Gyr), yet rotates with a
 period of only 19.8 $\pm$ 1.3 days.
-This star is relatively metal poor: $\mathrm{[Fe/H]} = -0.13 \pm$ 0.09 dex,
-and that could be the cause of its anomalously rapid rotation.
+This is KIC 9098294, a single-lined spectroscopic binary with an orbital
+period of around 20 days (Latham, private communication).
+It is the only clear SB1 in the \citet{vansaders2016} sample, although some
+others do have binary companions with long orbital periods, for which tidal
+interactions are not expected to be strong.
 }
 
 \begin{figure}
