### 2019-08-09 d6a5559
some small changes before resubmitting
