+++ paper/ms.tex
@@ -75,7 +75,8 @@
 \newcommand{\eep}{equivalent evolutionary point}
 \newcommand{\Eep}{Equivalent evolutionary point}
 
-\newcommand{\racomment}[1]{{\color{blue}#1}}
+% \newcommand{\racomment}[1]{{\color{blue}#1}}
+\newcommand{\racomment}[1]{{\bf #1}}
 
 \begin{document}
 
