+++ paper/conclusion.tex
@@ -27,7 +27,7 @@ combined with age and metallicity, determines the mass, radius and \logg\ of a
 star, this means rotation periods can improve the precision of {\it all}
 stellar parameters.
 Although V-band extinction is marginalized over during inference, correcting
-photometry for dust-extinction {\it before} analysis, or including it as a
+photometry for dust-extinction before analysis, or including it as a
 prior can improve the accuracy of stellar ages measured with \sd.
 \racomment{
 We also tested \sd\ on a set of 21 \kepler\ asteroseismic stars
