+++ paper/method.tex
@@ -440,7 +440,7 @@ single sample from the posterior PDF.
 Ages were inferred with \sd\ using Markov Chain Monte Carlo.
 The joint posterior PDF over age, mass, metallicity, distance and extinction
 was sampled using the affine invariant ensemble sampler, {\tt emcee}
-\citep{foreman-mackey2013} with 24 walkers.
+\citep{foreman-mackey2013} with 50 walkers.
 Samples were drawn from the posterior PDF until 100 {\it independent} samples
 were obtained.
 We actively estimated the autocorrelation length, which indicates how many
