+++ paper/results.tex
@@ -551,11 +551,11 @@ Colored circles show ages inferred using isochrone fitting and gyrochronology
 combined via the \sd\ software package.
 Black triangles show the ages of all stars inferred via gyrochronology only
     and white circles show ages inferred via isochrone-fitting only.
-This plot shows that, for the majority of stars in this sample, isochrone
-    fitting dominates the age information and gyrochronology only contributes
-    significantly for two or three stars.
-This is a consequence of many of these stars being hot old, extremely metal
-    rich, extremely metal poor, or evolved.
+% This plot shows that, for the majority of stars in this sample, isochrone
+%     fitting dominates the age information and gyrochronology only contributes
+%     significantly for two or three stars.
+% This is a consequence of many of these stars being hot old, extremely metal
+%     rich, extremely metal poor, or evolved.
   }
   \centering
     \includegraphics[width=1.2\textwidth]{asteroseismic_results_nosdss_gaia}
