+++ paper/master.tex
@@ -1,0 +1,1553 @@
+% zip ms_files.zip ms.tex abstract.tex intro.tex method.tex results.tex
+% acknowledgments.tex appendix.tex
+% conclusion.tex Praesepe.pdf variance.pdf simulated_CMD.pdf
+% rotation_model_praesepe.pdf simulation_results.pdf NGC6819.pdf
+% NGC6819_results.pdf hz.bib asteroseismic_results_nosdss_gaia.pdf
+
+\documentclass[useAMS, usenatbib, preprint, 12pt]{aastex}
+% \documentclass[a4paper,fleqn,usenatbib,useAMS]{mnras}
+\usepackage{cite, natbib}
+\usepackage{float}
+\usepackage{amsmath}
+\usepackage{hyperref}
+\usepackage{epsfig}
+\usepackage{cases}
+\usepackage[section]{placeins}
+\usepackage{graphicx, subfigure}
+\usepackage{color}
+\usepackage{bm}
+
+\AtBeginDocument{\let\textlabel\label}
+
+\newcommand{\ie}{{\it i.e.}}
+\newcommand{\eg}{{\it e.g.}}
+\newcommand{\etal}{{\it et al.}}
+
+\newcommand{\kepler}{{\it Kepler}}
+\newcommand{\Kepler}{{\it Kepler}}
+\newcommand{\corot}{{\it CoRoT}}
+\newcommand{\Ktwo}{{\it K2}}
+\newcommand{\ktwo}{\Ktwo}
+\newcommand{\TESS}{{\it TESS}}
+\newcommand{\tess}{{\it TESS}}
+\newcommand{\LSST}{{\it LSST}}
+\newcommand{\lsst}{{\it LSST}}
+\newcommand{\Wfirst}{{\it WFIRST}}
+\newcommand{\wfirst}{{\it WFIRST}}
+\newcommand{\SDSS}{{\it SDSS}}
+\newcommand{\PLATO}{{\it PLATO}}
+\newcommand{\plato}{{\it PLATO}}
+\newcommand{\Gaia}{{\it Gaia}}
+\newcommand{\gaia}{{\it Gaia}}
+\newcommand{\panstarrs}{{\it PanSTARRS}}
+
+\newcommand{\Teff}{$T_{\mathrm{eff}}$}
+\newcommand{\teff}{$T_{\mathrm{eff}}$}
+\newcommand{\FeH}{[Fe/H]}
+\newcommand{\feh}{[Fe/H]}
+\newcommand{\prot}{$P_{\mathrm{rot}}$}
+\newcommand{\pmega}{$\bar{\pi}$}
+\newcommand{\mj}{$m_j$}
+\newcommand{\mh}{$m_h$}
+\newcommand{\mk}{$m_k$}
+\newcommand{\mx}{$m_x$}
+\newcommand{\logg}{log(g)}
+\newcommand{\dnu}{$\Delta \nu$}
+\newcommand{\numax}{$\nu_{\mathrm{max}}$}
+
+% \newcommand{\nsim_stars}{841}
+
+\newcommand{\amnh}{1}
+\newcommand{\cca}{2}
+\newcommand{\florida}{3}
+\newcommand{\hawaii}{4}
+\newcommand{\columbia}{5}
+\newcommand{\riverside}{6}
+\newcommand{\cuny}{7}
+\newcommand{\nyu}{8}
+\newcommand{\cds}{9}
+\newcommand{\mpia}{10}
+\newcommand{\yale}{11}
+
+\newcommand{\sd}{{\tt stardate}}
+\newcommand{\MS}{main sequence}
+\newcommand{\ms}{main sequence}
[…]
