### 2019-08-11 a57aa17
shortening abstract
