+++ paper/abstract.tex
@@ -1,28 +1,26 @@
-We present a new age-dating technique that combines gyrochronology and
+We present a new age-dating technique that combines gyrochronology with
 isochrone fitting to infer ages for FGKM main-sequence and subgiant field
 stars.
-The age-dating methods of gyrochronology and isochrone fitting are each
-capable of providing relatively precise ages in certain areas of the
-Hertzsprung-Russell diagram: rotation periods can provide precise ages for
-cool stars on the main sequence via gyrochronology, and isochrone fitting
-can provide precise ages for stars near the main-sequence turnoff.
-Combined, these two age-dating techniques can be applied to a broader range of
-stellar masses and evolutionary stages and can provide ages that are more
-precise and accurate than either method used in isolation.
-In this investigation, we demonstrate that the position of a star on the
-Hertzsprung-Russell or color-magnitude diagram can be combined with its
-rotation period to infer a precise age via both isochrone fitting and
-gyrochronology simultaneously.
+Gyrochronology and isochrone fitting are each capable of providing relatively
+    precise ages for field stars in certain areas of the Hertzsprung-Russell
+    diagram: gyrochronology works optimally for cool main-sequence stars, and
+    isochrone fitting can provide precise ages for stars near the
+    main-sequence turnoff.
+Combined, these two age-dating techniques can provide precise and accurate
+ages for a broader range of stellar masses and evolutionary stages than either
+method used in isolation.
+We demonstrate that the position of a star on the Hertzsprung-Russell or
+color-magnitude diagram can be combined with its rotation period to infer a
+precise age via both isochrone fitting and gyrochronology simultaneously.
 We show that incorporating rotation periods with 5\% uncertainties into
-stellar evolution models particularly improves age precision for FGK stars on
-the main sequence, and can, on average, provide age estimates that are up to
-three times more precise than isochrone fitting alone for these stars.
+    stellar evolution models improves age precision for FGK stars on the main
+    sequence, and can, on average, provide age estimates up to three times
+    more precise than isochrone fitting alone.
 In addition, we provide a new gyrochronology relation, calibrated to the
 Praesepe cluster and the Sun, that includes a variance model to capture the
 rotational behavior of stars whose rotation periods do not lengthen with the
 square-root of time, and parts of the Hertzsprung-Russell diagram where
 gyrochronology has not been calibrated.
-This publication is accompanied by an open source {\it Python} package (\sd:
-\url{ https://github.com/ruthangus/stardate }) for inferring the ages of main
-sequence and subgiant FGKM stars from rotation periods, spectroscopic
-parameters and/or apparent magnitudes and parallaxes.
+This publication is accompanied by an open source {\it Python} package (\sd)
+for inferring the ages of main-sequence and subgiant FGKM stars from rotation
+periods, spectroscopic parameters and/or apparent magnitudes and parallaxes.
