+++ paper/ms.tex
@@ -1,4 +1,8 @@
-% zip ms_files.zip ms.tex abstract.tex intro.tex method.tex results.tex conclusion.tex Praesepe.pdf variance.pdf simulated_CMD.pdf rotation_model_praesepe.pdf simulation_results.pdf NGC6819.pdf NGC6819_results.pdf hz.bib
+% zip ms_files.zip ms.tex abstract.tex intro.tex method.tex results.tex
+% acknowledgments.tex appendix.tex
+% conclusion.tex Praesepe.pdf variance.pdf simulated_CMD.pdf
+% rotation_model_praesepe.pdf simulation_results.pdf NGC6819.pdf
+% NGC6819_results.pdf hz.bib asteroseismic_results_nosdss_gaia.pdf
 
 \documentclass[useAMS, usenatbib, preprint, 12pt]{aastex}
 % \documentclass[a4paper,fleqn,usenatbib,useAMS]{mnras}
