### 2019-08-20 cfe38f2
correcting a few typos before posting to ArXiv
