+++ paper/ms.tex
@@ -80,14 +80,10 @@
 \newcommand{\Eep}{Equivalent evolutionary point}
 
 % \newcommand{\racomment}[1]{{\color{blue}#1}}
-\newcommand{\racomment}[1]{{\bf #1}}
+% \newcommand{\racomment}[1]{{\bf #1}}
 
 \begin{document}
 
-% \title{Inferring stellar ages by combining isochrone fitting with
-% gyrochronology}
-% \title{A new age-dating model for cool main sequence stars that combines
-% isochrone fitting with gyrochronology}
 \title{Towards precise stellar ages: combining isochrone fitting with
 empirical gyrochronology}
 
