+++ paper/master.tex
@@ -153,7 +153,7 @@ Praesepe cluster and the Sun, that includes a variance model to capture the
 rotational behavior of stars whose rotation periods do not lengthen with the
 square-root of time, and parts of the Hertzsprung-Russell diagram where
 gyrochronology has not been calibrated.
-This publication is accompanied by an open source {\it Python} package (\sd)
+This publication is accompanied by an open source {\it Python} package, \sd,
 for inferring the ages of main-sequence and subgiant FGKM stars from rotation
 periods, spectroscopic parameters and/or apparent magnitudes and parallaxes.
 \end{abstract}
@@ -173,7 +173,7 @@ diagram (HRD) or color-magnitude diagram (CMD).
 On the MS, isochrones are tightly spaced and, even with very precise
 measurements of effective temperature and luminosity, the position of a MS
 star on the HRD may be consistent with range of isochrones spanning several
-billion years.
+billion years \citep[see][for a review of stellar ages]{soderblom2010}.
 At the main-sequence turnoff however, isochrones are spread further apart, so
 that sufficiently precisely measured temperatures and luminosities can yield
 ages that are extremely precise (with minimum statistical uncertainties on the
@@ -389,12 +389,14 @@ line in figure \ref{fig:praesepe}).
 In order to fit a relation to Praesepe, we removed rotational outliers bluer
 than \gcolor\ = 2.7 via sigma-clipping and fit a 5th-order polynomial to the
 remaining FGK and early M stars.
-\racomment{We found that a 5th order polynomial provided a substantially
+% \racomment{
+    We found that a 5th order polynomial provided a substantially
 better fit than lower-order polynomials, which were not able to capture the
 sharp `elbow’ in the rotation period-color relation. Additional orders
 provided either a worse
 fit, tending towards extreme values at the boundaries, or diminishing returns
-in goodness-of-fit.}
+in goodness-of-fit.
+% }
 We also fit a straight line to the late M dwarfs (\gcolor\ $>$ 2.7), to
 capture the mass-dependent initial rotation periods of low mass stars
 \citep{somers2017}.
@@ -812,7 +814,7 @@ extinctions at random for 1000 stars from the following uniform distributions:
 & D \sim U(10, 1000)~\mathrm{[pc]} \\
 & A_V \sim U(0, 0.1).
 \end{eqnarray}
-\teff, \logg, \fhat, parallax, and apparent magnitudes $B$, $V$, $J$, $H$, $K$,
+\teff, \logg, [Fe/H], parallax, and apparent magnitudes $B$, $V$, $J$, $H$, $K$,
 \gaia\ $G$, $G_{BP}$ and $G_{RP}$ were generated from these
 stellar parameters using the MIST stellar evolution models.
 We added a small amount of noise to the `observed' stellar properties in order
@@ -969,7 +971,7 @@ Neither gyrochronology nor isochrone fitting can provide precise ages for
 
 An important caveat associated with these results is that they strongly depend
 on the uncertainties adopted for all observables used in the analysis: \teff,
-\feh, \logg, $G$, $G_BP$, $G_RP$, $J$, $H$, $K$, $B$, $V$, and rotation
+\feh, \logg, $G$, $G_{BP}$, $G_{RP}$, $J$, $H$, $K$, $B$, $V$, and rotation
 period.
 Changing the uncertainties on these observables will affect the uncertainties
 on inferred ages in different ways.
@@ -1101,7 +1103,7 @@ Figure \ref{fig:NGC6819} shows the period-color relation of this
 cluster and figure \ref{fig:NGC6819_results} shows the results of inferring
 the ages of individual cluster members using a combination of gyrochronology
[…]
