### 2019-08-30 #1469 @arfon
Thanks @crawfordsm. @RuthAngus - can you merge this PR which adds the ApJ link to the paper? https://github.com/RuthAngus/stardate/pull/8.

Also, I forgot to check, is it ApJ you are publishing the accompanying paper in? If not, just modify the name of the journal on [this line](https://github.com/RuthAngus/stardate/pull/8/files#diff-e2778e7674f45b806282a9611faa7220R24).
