### 2019-09-03 0ab5c8d
adding AAS DOI and journal
