### 2019-09-03 #1469 @RuthAngus
Hi @arfon and @crawfordsm, it's an AJ paper, so I added the DOI and journal lines to the paper file without merging the PR, hope that's ok!
