### 2019-09-03 #1469 @arfon
@RuthAngus - do you think you could update the reference to 'Angus et al (submitted)' to point to the AJ paper (`https://doi.org/10.3847/1538-3881/ab3c53`)?
