### 2019-09-13 16354d5
setting save_samples=False by default
