### 2019-09-13 fc8d9ef
fixing bug in save_samples
