### 2019-09-13 184727b
updating for save_samples=False by default
