### 2019-09-16 ba6ecbd
updating reference to AJ paper
