+++ paper/references.bib
@@ -64,3 +64,15 @@ archivePrefix = "arXiv",
    adsurl = {http://adsabs.harvard.edu/abs/2016ApJ...823..102C},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{angus2019,
+   author = {{Angus}, R. and {Morton}, T. D. and {Foreman-Mackey}, D. and {van Saders}, J. and
+	{Curtis}, J. L. and {Kane}, S.~R. and {Bedell}, M. and {Kiman}, R. {Hogg}, D. W. and {Brewer}, J.},
+    title = "{Towards precise stellar ages: combining isochrone fitting with empirical gyrochronology}",
+  journal = {\aj},
+archivePrefix = "arXiv",
+   eprint = {1908.07528},
+ primaryClass = "astro-ph.SR",
+     year = 2019,
+      doi = {https://doi.org/10.3847/1538-3881/ab3c53},
+}
