### 2019-09-18 #1469 @RuthAngus
Hi @arfon, I have updated the reference in the paper and created a new release! The Zenodo DOI is: 10.5281/zenodo.3444905
